+++ src/page-layer/signin/SignIn.tsx
@@ -2,56 +2,11 @@
 
 import Image from "next/image";
 
-import { Button } from "@/shared/ui";
+import { CheckIcon } from "@/shared/assets";
 
+import { googleLogin } from "./api/hooks";
+import { features } from "./model/constants";
 import styles from "./SignIn.module.scss";
-
-const CheckIcon = () => (
-  <svg
-    width="20"
-    height="20"
-    viewBox="0 0 20 20"
-    fill="none"
-    xmlns="http://www.w3.org/2000/svg"
-  >
-    <circle cx="10" cy="10" r="10" fill="#048724" />
-    <path
-      d="M6 10L9 13L14 7"
-      stroke="white"
-      strokeWidth="2"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-);
-
-const GoogleIcon = () => (
-  <svg width="20" height="20" viewBox="0 0 20 20" fill="none">
-    <path
-      d="M19.6 10.2273C19.6 9.51818 19.5364 8.83636 19.4182 8.18182H10V12.05H15.3818C15.15 13.3 14.4455 14.3591 13.3864 15.0682V17.5773H16.6182C18.5091 15.8364 19.6 13.2727 19.6 10.2273Z"
-      fill="#4285F4"
-    />
-    <path
-      d="M10 20C12.7 20 14.9636 19.1045 16.6182 17.5773L13.3864 15.0682C12.4909 15.6682 11.3455 16.0227 10 16.0227C7.39545 16.0227 5.19091 14.2636 4.40455 11.9H1.06364V14.4909C2.70909 17.7591 6.09091 20 10 20Z"
-      fill="#34A853"
-    />
-    <path
-      d="M4.40455 11.9C4.20455 11.3 4.09091 10.6591 4.09091 10C4.09091 9.34091 4.20455 8.7 4.40455 8.1V5.50909H1.06364C0.386364 6.85909 0 8.38636 0 10C0 11.6136 0.386364 13.1409 1.06364 14.4909L4.40455 11.9Z"
-      fill="#FBBC05"
-    />
-    <path
-      d="M10 3.97727C11.4682 3.97727 12.7864 4.48182 13.8227 5.47273L16.6909 2.60455C14.9591 0.990909 12.6955 0 10 0C6.09091 0 2.70909 2.24091 1.06364 5.50909L4.40455 8.1C5.19091 5.73636 7.39545 3.97727 10 3.97727Z"
-      fill="#EA4335"
-    />
-  </svg>
-);
-
-const features = [
-  "일일 출석 체크",
-  "부트캠프 진행 상황 모니터링",
-  "훈련 수당 및 공제 내역 조회",
-  "휴가 일수 및 신청 관리",
-];
 
 const SignIn = () => {
   return (
@@ -59,7 +14,7 @@ const SignIn = () => {
       <div className={styles.card}>
         <div className={styles.header}>
           <div className={styles.icon}>
-            <Image alt="logo" src="/logo.svg" width={64} height={64} priority />
+            <Image alt="logo" src="/logo.svg" width={80} height={80} priority />
           </div>
           <h1 className={styles.title}>PLANIT</h1>
           <p className={styles.subtitle}>
@@ -76,15 +31,15 @@ const SignIn = () => {
           ))}
         </div>
 
-        <div className={styles.buttonContainer}>
-          <Button
-            variant="outline"
-            width="100%"
-            icon={<GoogleIcon />}
-            className={styles.googleButton}
-          >
-            Google로 계속하기
-          </Button>
+        <div className={styles.buttonContainer} onClick={googleLogin}>
+          <Image
+            src="/google.svg"
+            width={200}
+            height={50}
+            alt="google login"
+            className={styles.googleImg}
+            priority
+          />
         </div>
       </div>
     </div>
