+++ src/page-layer/signin/model/constants.ts
@@ -1,0 +1,6 @@
+export const features = [
+  "일일 출석 체크",
+  "부트캠프 진행 상황 모니터링",
+  "훈련 수당 및 공제 내역 조회",
+  "휴가 일수 및 신청 관리",
+];
